task 2

+++ 0x14-bit_manipulation/2-get_bit.c
@@ -6,7 +6,7 @@
  * @index: index
  * Return: the bit
  */
-int get_bit(unsigned int n, unsigned int index)
+int get_bit(unsigned long int n, unsigned int index)
 {
 	if (index >= (sizeof(unsigned long int) * 8))
 		return (-1);
